Drop concept of copy-modes
This was an idiotic idea to start with. It deviates from the traditional
user-centric approach expected of Atom packages, and the better solution
is to define separate commands to copy source for different languages. A
user can then assign whatever keybindings they wish to "copy-regexp-js",
"copy-regexp-cson", etc.

More importantly, consuming the status-bar and constructing views adds a
major hit to load time (~120ms, which is *far* too long for a language's
package to initialise).