Update version string in package.json
This isn't a release. Nothing's been changed since v1.1.2, except that I
attempted to correct a mistake too trivial to warrant mention, believing
I could republish the same tag to APM after it was recently unpublished.
Evidently, you can't.

I therefore refuse to acknowledge this as a "release", yet I'm forced to
tag it incrementally on GitHub in order for Atom's package repository to
accept it. Because of these bullshit, stringent limitations, and because
I'm a stubborn bastard, I'm refusing to conform quietly. This might also
be the last time I publish a release to atom.io that isn't `file-icons`.